requirements.txt
streamlit==1.51.0
groq>=0.10.0
easyocr==1.7.2
pymupdf==1.26.7
python-dotenv==1.2.1
fpdf2==2.8.5
faker>=22.0.0
cryptography>=41.0.0
pillow>=10.0.0
numpy>=1.24.0

